+++ docs/developer-docs/docusaurus.config.ts
@@ -19,7 +19,6 @@ const config: Config = {
     projectName: "wardenprotocol", // Usually your repo name.
 
     onBrokenLinks: "throw",
-    onBrokenMarkdownLinks: "warn",
 
     // Even if you don't use internationalization, you can use this field to set
     // useful metadata like html lang. For example, if your site is Chinese, you
@@ -31,6 +30,11 @@ const config: Config = {
 
     markdown: {
         mermaid: true,
+    },
+    customFields: {
+        markdownHooks: {
+            onBrokenMarkdownLinks: "warn",
+        },
     },
     themes: ["@docusaurus/theme-mermaid"],